introduced cleanOutdatedUser cron

+++ cloud/functions/src/functions/firestore/services/user-utils.ts
@@ -25,10 +25,12 @@ export async function getUserDocsMatching(collectionFilter: (collection: Collect
 export async function cleanOutdatedUsers() {
   const oldestValidLastConnectionDate = Temporal.Now.zonedDateTimeISO().subtract({months: 2}).toString() as ISODatetime
   const MAXIMUM_NUMBER_OF_FAVS = 0;
+  const MAXIMUM_NUMBER_OF_FEEDBACKS = 0;
 
   const stats = await windowedProcessUsers(
     () => db.collection(`/users`)
       .where(`totalFavs.total`, '<=', MAXIMUM_NUMBER_OF_FAVS)
+      .where(`totalFeedbacks.total`, '<=', MAXIMUM_NUMBER_OF_FEEDBACKS)
       .where(`userLastConnection`, '<', oldestValidLastConnectionDate),
     (userDoc) => deleteUserRefIncludingChildren(userDoc.ref),
   )
@@ -110,8 +112,73 @@ export async function windowedProcessUsers(
   return stats;
 }
 
-async function deleteUserRefIncludingChildren(userRef: DocumentReference<DocumentData>) {
-  await userRef.delete()
+async function deleteUserRefIncludingChildren(userRef: DocumentReference<DocumentData>, force = false) {
+  const [preferencesDoc, tokensWalletDoc] = await Promise.all([
+    db.doc(`/users/${userRef.id}/preferences/self`).get(),
+    db.doc(`/users/${userRef.id}/tokens-wallet/self`).get(),
+  ])
 
-  // TODO: delete favs & feedbacks sub collections (don't forget space-based data)
+  // let's not delete
+  const collectionsPreventingDelete = ([] as string[])
+    .concat(preferencesDoc.exists ? ["preferences"]:[])
+    .concat(tokensWalletDoc.exists ? ["tokens-wallet"]:[])
+  if(collectionsPreventingDelete.length && !force) {
+    console.info(`Not deleting user ${userRef.id} because he has ${collectionsPreventingDelete.join("/")} non-empty collection`)
+    return;
+  }
+
+  await Promise.all([
+    deleteUserSpaces(userRef),
+    deleteUserEventsFromNode(userRef),
+  ])
+
+  console.info(`Deleting user entry ${userRef.path}`)
+  await userRef.delete()
 }
+
+async function deleteUserEventsFromNode(rootRef: DocumentReference) {
+  const eventRefs = await db.collection(`${rootRef.path}/events`).listDocuments();
+
+  await Promise.all(eventRefs.map(async eventRef => {
+    console.debug(`Deleting user event entry ${eventRef.path}...`)
+    return Promise.all([
+      deleteUserTalkNotesFromEvent(eventRef),
+      deleteUserDailyFeedbacksFromEvent(eventRef),
+      eventRef.delete(),
+    ])
+  }))
+}
+
+async function deleteUserSpaces(userRef: DocumentReference<DocumentData>) {
+  const spaceRefs = await db.collection(`${userRef.path}/spaces`).listDocuments()
+
+  await Promise.all(spaceRefs.map(async spaceRef => {
+    console.debug(`Deleting user space entry ${spaceRef.path}...`)
+    return Promise.all([
+      deleteUserEventsFromNode(spaceRef),
+      spaceRef.delete(),
+    ])
+  }))
+}
+
+async function deleteUserTalkNotesFromEvent(eventRef: DocumentReference<DocumentData>) {
+  const talkNotesRefs = await db.collection(`${eventRef.path}/talksNotes`).listDocuments()
+
+  await Promise.all(talkNotesRefs.map(async talkNotesRef => {
+    console.debug(`Deleting user talk notes entry ${talkNotesRef.path}...`)
+    return talkNotesRef.delete()
+  }))
+}
+
+async function deleteUserDailyFeedbacksFromEvent(eventRef: DocumentReference<DocumentData>) {
+  const dailyFeedbackRefs = await db.collection(`${eventRef.path}/days`).listDocuments()
+
+  await Promise.all(dailyFeedbackRefs.map(async dailyFeedbackRef => {
+    console.debug(`Deleting user daily feedback entry ${dailyFeedbackRef.path}...`)
+    return Promise.all([
+      db.doc(`${dailyFeedbackRef.path}/feedbacks/self`).delete(),
+      dailyFeedbackRef.delete(),
+    ])
+  }))
+}
+
